Fix naming pattern
