+++ ui/test/nuxeo-move-documents-up-button.test.js
@@ -16,12 +16,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 import { fixture, html, login, tap, waitForEvent } from '@nuxeo/testing-helpers';
-import '../actions/nuxeo-move-documents-down-button.js';
 import '../actions/nuxeo-move-documents-up-button.js';
 
 suite('Given on a Ordered Folder, I have 5 items', () => {
   let server;
-  let downButton;
   let upButton;
   const items = [{ uid: '1' }, { uid: '2' }, { uid: '3' }, { uid: '4' }, { uid: '5' }];
   setup(async () => {
@@ -33,27 +31,13 @@ suite('Given on a Ordered Folder, I have 5 items', () => {
     ]);
   });
 
-  suite('When I select items 1, 3, 4, 5', async () => {
+  suite('When I select items 1, 3, 4, 5', () => {
     const selectedItems = [items[0], items[2], items[3], items[4]];
     setup(async () => {
-      downButton = await fixture(html`
-        <nuxeo-move-documents-down-button .documents=${items.slice(0)} .selectedDocuments=${selectedItems.slice(0)}>
-        </nuxeo-move-documents-down-button>
-      `);
       upButton = await fixture(html`
         <nuxeo-move-documents-up-button .documents=${items.slice(0)} .selectedDocuments=${selectedItems.slice(0)}>
         </nuxeo-move-documents-up-button>
       `);
-    });
-
-    test('Then clicking "Down" should order items to 2, 3, 4, 5', async () => {
-      tap(downButton);
-      await waitForEvent(downButton, 'refresh-display');
-      assert.equal(2, downButton.documents[0].uid);
-      assert.equal(1, downButton.documents[1].uid);
-      assert.equal(3, downButton.documents[2].uid);
-      assert.equal(4, downButton.documents[3].uid);
-      assert.equal(5, downButton.documents[4].uid);
     });
 
     test('Then clicking "Up" should order items to 1, 2, 3, 4', async () => {
@@ -87,48 +71,13 @@ suite('Given on a Ordered Folder, I have 5 items', () => {
     });
   });
 
-  suite('When I select items 2, 4, 5', () => {
-    const selectedItems = [items[1], items[3], items[4]];
-    setup(async () => {
-      downButton = await fixture(html`
-        <nuxeo-move-documents-down-button .documents=${items.slice(0)} .selectedDocuments=${selectedItems.slice(0)}>
-        </nuxeo-move-documents-down-button>
-      `);
-    });
-
-    test('Then clicking "Down" should order items to 3, 4, 5', () => {
-      tap(downButton);
-      return waitForEvent(downButton, 'refresh-display').then(() => {
-        assert.equal(1, downButton.documents[0].uid);
-        assert.equal(3, downButton.documents[1].uid);
-        assert.equal(2, downButton.documents[2].uid);
-        assert.equal(4, downButton.documents[3].uid);
-        assert.equal(5, downButton.documents[4].uid);
-      });
-    });
-  });
-
   suite('When I select items 2, 4', () => {
     const selectedItems = [items[1], items[3]];
     setup(async () => {
-      downButton = await fixture(html`
-        <nuxeo-move-documents-down-button .documents=${items.slice(0)} .selectedDocuments=${selectedItems.slice(0)}>
-        </nuxeo-move-documents-down-button>
-      `);
       upButton = await fixture(html`
         <nuxeo-move-documents-up-button .documents=${items.slice(0)} .selectedDocuments=${selectedItems.slice(0)}>
         </nuxeo-move-documents-up-button>
       `);
-    });
-
-    test('Then clicking "Down" should order items to 4, 5', async () => {
-      tap(downButton);
-      await waitForEvent(downButton, 'refresh-display');
-      assert.equal(1, downButton.documents[0].uid);
-      assert.equal(3, downButton.documents[1].uid);
-      assert.equal(5, downButton.documents[2].uid);
-      assert.equal(2, downButton.documents[3].uid);
-      assert.equal(4, downButton.documents[4].uid);
     });
 
     test('Then clicking "Up" should order items to 1, 2', async () => {
@@ -160,18 +109,53 @@ suite('Given on a Ordered Folder, I have 5 items', () => {
       assert.equal(false, upButton._available);
     });
   });
+});
 
-  suite('When I select items 4, 5', () => {
-    const selectedItems = [items[3], items[4]];
-    setup(async () => {
-      downButton = await fixture(html`
-        <nuxeo-move-documents-down-button .documents=${items.slice(0)} .selectedDocuments=${selectedItems}>
-        </nuxeo-move-documents-down-button>
-      `);
+suite('nuxeo-move-documents-up-button extras', () => {
+  let el;
+
+  setup(async () => {
+    el = await fixture(
+      html`
+        <nuxeo-move-documents-up-button></nuxeo-move-documents-up-button>
+      `,
+    );
+  });
+
+  suite('_isAvailable', () => {
+    test('sets available=false initially', () => {
+      el.selectedDocuments = null;
+      el._isAvailable();
+      expect(el._available).to.be.false;
     });
 
-    test('Then I cannot click "Down"', () => {
-      assert.equal(false, downButton._available);
+    test('sets available=false for empty', () => {
+      el.selectedDocuments = [];
+      el._isAvailable();
+      expect(el._available).to.be.false;
+    });
+
+    test('dispatches clear-selected-items on sort error', (done) => {
+      el.addEventListener('clear-selected-items', () => done());
+      el.documents = [{ uid: 'a' }];
+      el.selectedDocuments = [{ uid: 'b' }, { uid: 'c' }];
+      el._isAvailable();
+    });
+
+    test('sets available=true for valid non-top selection', () => {
+      const docs = [{ uid: '1' }, { uid: '2' }, { uid: '3' }];
+      el.documents = docs;
+      el.selectedDocuments = [docs[2]];
+      el._isAvailable();
+      expect(el._available).to.be.true;
+    });
+
+    test('returns early for top doc that is a sequence', () => {
+      const docs = [{ uid: '1' }, { uid: '2' }];
+      el.documents = docs;
+      el.selectedDocuments = [docs[0]];
+      el._isAvailable();
+      expect(el._available).to.be.false;
     });
   });
 });
